Extract method and add unit test
